Cross-Section Module

Reading HITRAN Files

RadiativeTransfer.CrossSection.readHITRAN — Function

readHITRAN(filename::String, molecule::Int, isotope::Int, ν_min::Number, ν_max::Number)

Read/parse a HITRAN data file and return the data in HitranTable format

Computing Cross Sections

Types

RadiativeTransfer.CrossSection.AbstractCrossSection — Type

type AbstractCrossSection

Abstract Cross Section type for generic cross section calculations

RadiativeTransfer.CrossSection.HitranTable — Type

struct HitranCrossSection{FT}

An AbstractCrossSection type struct, which provides all HITRAN line parameters needed to compute absorption cross sections

See https://hitran.org/docs/definitions-and-units/ for details

Fields

mol

The molecular species identification (ID) number

iso

The isotopologue ID number

νᵢ

The wavenumber of the spectral line transition (cm-1) in vacuum

Sᵢ

The spectral line intensity (cm−1/(molecule·cm−2)) at Tref=296K

Aᵢ

The Einstein-A coefficient (s-1) of a transition

γ_air

The air-broadened half width at half maximum (HWHM) (cm−1/atm) at Tref=296K and reference pressure pref=1atm

γ_self

The self-broadened half width at half maximum (HWHM) (cm−1/atm) at Tref=296K and reference pressure pref=1atm

E″

The lower-state energy of the transition (cm-1)

n_air

The coefficient of the temperature dependence of the air-broadened half width

δ_air

The pressure shift (cm−1/atm) at Tref=296K and pref=1atm of the line position with respect to the vacuum transition wavenumber νij

global_upper_quanta

The electronic and vibrational quantum numbers and labels for the upper state of a transition

global_lower_quanta

The electronic and vibrational quantum numbers and labels for the lower state of a transition

local_upper_quanta

Rotational, hyperfine and other quantum numbers and labels for the upper state of a transition

local_lower_quanta

Rotational, hyperfine and other quantum numbers and labels for the lower state of a transition

ierr

Ordered list of indices corresponding to uncertainty estimates of transition parameters

iref

Ordered list of reference identifiers for transition parameters

line_mixing_flag

A flag indicating the presence of additional data and code relating to line-mixing

g′

The upper state degeneracy

g″

The lower state degeneracy