Cross-Section Module

Introduction

This module enables absorption cross-section calculations of atmospheric gases at different pressures, temperatures, and broadeners (Doppler, Lorentzian, Voigt). It uses the HITRAN energy transition database for calculations. While it enables lineshape calculations from scratch, it also allows users to create and save an interpolator object at specified wavelength, pressure, and temperature grids. It can perform these computations either on CPU or GPU (CUDA).

You can calculate an absorption cross-section in three steps:

Use read_hitran to read in a HITRAN database file

Use make_hitran_model or make_interpolation_model to set up model parameters for the cross-section calculation

Use absorption_cross_section to perform the cross-section calculation using the defined model settings

Example

using RadiativeTransfer
using RadiativeTransfer.CrossSection

## 
## STEP 1: Get the Hitran data
## 

# If you have a fixed-width file stored locally
hitran_data = read_hitran("path/to/file", mol=2, iso=1, ν_min=6000, ν_max=6400)

# If you would like to download the file from the hitran database
hitran_data = read_hitran(artifact"hitran_molec_id_2_CO2", mol=2, iso=1, ν_min=6000, ν_max=6400)

## 
## STEP 2: Create a model from parameters
## 

# These are some example models, but you can create/customize your model however you'd like. 
# Please see make_hitran_model documentation for optional arguments
model_doppler = make_hitran_model(hitran_data, Doppler())
model_lorentz = make_hitran_model(hitran_data, Lorentz())
model_voigt_CPU = make_hitran_model(hitran_data, Voigt())
model_voigt_GPU = make_hitran_model(hitran_data, Voigt(), architecture=Architectures.GPU())

# If you would prefer to create an interpolation, and then interpolate the cross-section 
# at other wavelengths/pressures/temperatures, you can use make_interpolation_model as such:

ν_grid = 6000:0.01:6400
pressures = 250:250:1250
temperatures = 100:75:400

model_interp = make_interpolation_model(hitran_data, Voigt(), ν_grid, pressures, temperatures)

# Note: please see make_interpolation_model documentation for optional parameters

## 
## STEP 3: Calculate the absorption cross section with the created model
## 

# You can specify the wavelength grid, pressure, and temperature. 
absorption_cross_section(model_*, 6000:0.01:6400, 1000.1, 296.1)

Reading HITRAN Files

RadiativeTransfer.CrossSection.read_hitran — Function

read_hitran(filepath::String, mol::Int=-1, iso::Int=-1, ν_min::Real=0, ν_max::Real=Inf)

Read/parse a HITRAN data file and return the data in HitranTable format

Defining Models with Cross-Section Parameters

RadiativeTransfer.CrossSection.make_hitran_model — Function

make_hitran_model(hitran::HitranTable, 
                  broadening::AbstractBroadeningFunction; 
                  wing_cutoff::Real=40, 
                  vmr::Real=0, 
                  CEF::AbstractComplexErrorFunction=HumlicekWeidemann32SDErrorFunction(), 
                  architecture = default_architecture)

Convenience function to make a HitranModel out of the parameters (Matches makeinterpolationmodel)

RadiativeTransfer.CrossSection.make_interpolation_model — Function

make_interpolation_model(hitran::HitranTable, 
                broadening::AbstractBroadeningFunction, 
                wave_grid::AbstractRange{<:Real}, 
                p_grid::AbstractRange{<:Real},
                t_grid::AbstractRange{<:Real}; 
                wavelength_flag::Bool=false,
                wing_cutoff::Real=40, 
                vmr::Real=0, 
                CEF::AbstractComplexErrorFunction=HumlicekWeidemann32SDErrorFunction(),
                architecture::AbstractArchitecture=default_architecture )

Using a HitranModel, create an InterpolationModel by interpolating the lineshape function at the given pressure and temperature grids

Computing Absorption Cross-Sections

RadiativeTransfer.CrossSection.absorption_cross_section — Function

absorption_cross_section(model::HitranModel, grid::AbstractRange{<:Real}, pressure::Real, temperature::Real; wavelength_flag::Bool=false)

Given a HitranModel, return the calculated absorption cross-section at the given pressure, temperature, and grid of wavelengths (or wavenumbers)

absorption_cross_section(model::InterpolationModel, grid::AbstractRange{<:Real}, pressure::Real, temperature::Real; wavelength_flag::Bool=false)

Given an Interpolation Model, return the interpolated absorption cross-section at the given pressure, temperature, and grid of wavelengths (or wavenumbers)

absorption_cross_section(model::HitranModel, grid::AbstractRange{<:Real}, pressure::Real, temperature::Real; wavelength_flag::Bool=false)

Calculate absorption cross-section at the given pressure, temperature, and grid of wavelengths (or wavenumbers), and have the option to perform auto-differentiation

Types

Hitran Data Structure Type

RadiativeTransfer.CrossSection.HitranTable — Type

struct HitranTable{FT}

A struct, which provides all HITRAN line parameters needed to compute absorption cross sections

See https://hitran.org/docs/definitions-and-units/ for details

Fields

mol

The molecular species identification (ID) number

iso

The isotopologue ID number

νᵢ

The wavenumber of the spectral line transition (cm-1) in vacuum

Sᵢ

The spectral line intensity (cm−1/(molecule·cm−2)) at Tref=296K

Aᵢ

The Einstein-A coefficient (s-1) of a transition

γ_air

The air-broadened half width at half maximum (HWHM) (cm−1/atm) at Tref=296K and reference pressure pref=1atm

γ_self

The self-broadened half width at half maximum (HWHM) (cm−1/atm) at Tref=296K and reference pressure pref=1atm

E″

The lower-state energy of the transition (cm-1)

n_air

The coefficient of the temperature dependence of the air-broadened half width

δ_air

The pressure shift (cm−1/atm) at Tref=296K and pref=1atm of the line position with respect to the vacuum transition wavenumber νij

global_upper_quanta

The electronic and vibrational quantum numbers and labels for the upper state of a transition

global_lower_quanta

The electronic and vibrational quantum numbers and labels for the lower state of a transition

local_upper_quanta

Rotational, hyperfine and other quantum numbers and labels for the upper state of a transition

local_lower_quanta

Rotational, hyperfine and other quantum numbers and labels for the lower state of a transition

ierr

Ordered list of indices corresponding to uncertainty estimates of transition parameters

iref

Ordered list of reference identifiers for transition parameters

line_mixing_flag

A flag indicating the presence of additional data and code relating to line-mixing

g′

The upper state degeneracy

g″

The lower state degeneracy

Broadening Function Types

RadiativeTransfer.CrossSection.Doppler — Type

Doppler line broadening

RadiativeTransfer.CrossSection.Lorentz — Type

Lorentz line broadening

RadiativeTransfer.CrossSection.Voigt — Type

Voigt line broadening

Complex Error Function Types

RadiativeTransfer.CrossSection.HumlicekErrorFunction — Type

Humlicek only formulation for Complex Error Function

RadiativeTransfer.CrossSection.HumlicekWeidemann32VoigtErrorFunction — Type

Mix of Humlicek and Weidemann (N=32) Error Function, suggested for Voigt function

RadiativeTransfer.CrossSection.HumlicekWeidemann32SDErrorFunction — Type

Mix of Humlicek and Weidemann (N=32) Error Function, suggested for Speed Dependent Voigt function

RadiativeTransfer.CrossSection.CPF12ErrorFunction — Type

Humlicek with a single rational approximation.

RadiativeTransfer.CrossSection.ErfcHumliErrorFunctionVoigt — Type

Mix of Humplicek and erfc Special Function for Voigt

RadiativeTransfer.CrossSection.ErfcHumliErrorFunctionSD — Type

Mix of Humplicek and erfc Special Function for Voigt

RadiativeTransfer.CrossSection.ErfcErrorFunction — Type

erfc Special Function for Voigt

Cross-Section Model Types

RadiativeTransfer.CrossSection.HitranModel — Type

struct HitranModel{FT}

A struct which provides all model parameters needed for cross-section calculations using HITRAN data

Fields

hitran

Struct with hitran data

broadening

Broadening function (Doppler/Lorentz/Voigt)

wing_cutoff

Wing cutoff [cm-1]

vmr

VMR of gas itself [0-1]

CEF

Complex Error Function to Use

architecture

Computer Architecture on which Model is run

RadiativeTransfer.CrossSection.InterpolationModel — Type

struct InterpolationModel{FT}

A struct which provides all model parameters needed for cross-section calculations using an Interpolator

Fields

itp

The interpolator

mol

The molecular species identification (ID) number

iso

The isotopologue ID number

broadening

Broadening function (Doppler/Lorentz/Voigt), as a string

ν_grid

Wavelength grids

p_grid

Wavelength grids

t_grid

Wavelength grids

wing_cutoff

Wing cutoff [cm-1]

vmr

VMR of gas itself [0-1]

CEF

Complex Error Function to Use

architecture

Computer Architecture on which Model is run