RadiativeTransfer Module

Introduction

The RadiativeTransfer module allows end-to-end simulation of radiative transfer (RT) throughout Earth's atmosphere and surface. Specifically, it:

Enables 1D vectorized plane-parallel RT modeling based on the Matrix Operator Method.

Incorporates fast, high fidelity simulations of scattering atmospheres containing haze and clouds – including pressure- and temperature-resolved absorption profiles of gaseous species in the atmosphere.

Enables GPU-accelerated computations of the resulting hyperspectral reflectances/transmittances.

Example

using RadiativeTransfer
using RadiativeTransfer.vSmartMOM

## 
## STEP 1: Load / Customize RT parameters
## 

# If you would like to load your own parameters from a YAML file
# (See required format at: https://github.com/RupeshJey/RadiativeTransfer.jl/blob/master/src/vSmartMOM/ModelParameters/DefaultParameters.yaml)
parameters = parameters_from_yaml("RadiativeTransfer/src/vSmartMOM/ModelParameters/DefaultParameters.yaml")

# OR if you would like to load a default set of parameters
parameters = default_parameters();

# You can then change any individual fields in parameters (parameters.field = ...)

## 
## STEP 2: Create a model from parameters
## 

# Creating this model will use the parameters object to calculate derived fields and attributes
model = model_from_parameters(parameters);

# Again, you can then change any derived field in model (model.field = ...)

## 
## STEP 3: Run the RT simulation
## 

R = rt_run(model)

References

1. Sanghavi, S., Davis, A.B. and Eldering, A., 2014. vSmartMOM: A vector matrix operator 
method-based radiative transfer model linearized with respect to aerosol properties. Journal of 
Quantitative Spectroscopy and Radiative Transfer, 133, pp.412-433.

2. Sanghavi, S. and Stephens, G., 2015. Adaptation of the delta-m and δ-fit truncation methods 
to vector radiative transfer: Effect of truncation on radiative transfer accuracy. Journal of 
Quantitative Spectroscopy and Radiative Transfer, 159, pp.53-68.

3. Grant, I.P. and Hunt, G.E., 1969. Discrete space theory of radiative transfer I. 
Fundamentals. Proceedings of the Royal Society of London. A. Mathematical and Physical Sciences, 
313(1513), pp. 183-197.

4. Plass, G.N., Kattawar, G.W. and Catchings, F.E., 1973. Matrix operator theory of radiative 
transfer. 1: Rayleigh scattering. Applied Optics, 12(2), pp.314-329.

Defining Parameters for RT Simulation

RadiativeTransfer.vSmartMOM.parameters_from_yaml — Function

Given a path to a YAML parameters file, load all the parameters into a vSmartMOM_Parameters struct

RadiativeTransfer.vSmartMOM.default_parameters — Function

Generate default set of parameters for Radiative Transfer calculations (from ModelParameters/)

Using the Parameters Object to Create a Model

RadiativeTransfer.vSmartMOM.model_from_parameters — Function

Take the parameters specified in the vSmartMOMParameters struct, and calculate derived attributes into a vSmartMOMModel

Performing the RT Simulation

RadiativeTransfer.vSmartMOM.rt_run — Function

rt_run(pol_type, obs_geom::ObsGeometry, τRayl, τAer, quad_points::QuadPoints, max_m, aerosol_optics, greek_rayleigh, τ_abs, brdf, architecture::AbstractArchitecture)

Perform Radiative Transfer calculations using given parameters

rt_run(model::vSmartMOM_Model, i_band::Integer = -1)

Perform Radiative Transfer calculations using parameters passed in through the vSmartMOM_Model struct

Types

Parameters Type

RadiativeTransfer.vSmartMOM.vSmartMOM_Parameters — Type

struct vSmartMOM_Parameters

A struct which holds all initial model parameters (before any computations)

Fields

float_type

Float type to use in the RT (Float64/Float32)

λ_band

band center wavelength (nBand)

λ_ref

reference wavelength

depol

Depolarization value

l_trunc

Truncation length for legendre terms (scalar for now, can do nBand later)

Δ_angle

Exclusion angle for forward peak

max_m

Hard cutoff for maximum number of m iterations

polarization_type

Type of polarization (I/IQ/IQU/IQUV)

obs_alt

Altitude of observer [Pa]

sza

Solar zenith angle [deg]

vza

Viewing zenith angles [deg]

vaz

Viewing azimuthal angles [deg]

nAer

Number of aerosol species

τAer_ref

AOD at Reference wavelength (nAer)

μ

Log mean radius (nAer)

σ

Log stddev of radius (nAer)

r_max

Maximum radius

nquad_radius

Number of quadrature points for integration of size distribution

nᵣ

Real part of refractive index (nBand,nAer)

nᵢ

Imag part of refractive index (nBand,nAer)

p₀

Pressure peak Pa

σp

Pressure peak width Pa

file

Path to atmospheric profile file

timeIndex

Time index to retrieve from file

lat

Latitude of atmospheric profile

lon

Longitude of atmospheric profile

profile_reduction_n

Length of profile reduction

spec_grid_start

Starting wavelength for absorption grid (nBand)

spec_grid_end

Ending wavelength for absorption grid (nBand)

spec_grid_n

Number of spectral points in absorption grid (nBand)

molecules

Molecules to use for absorption calculations (nBand, nMolecules)

broadening_function

Type of broadening function (Doppler/Lorentz/Voigt)

wing_cutoff

Wing cutoff to use in cross-section calculation

CEF

Complex Error Function to use in Voigt calculations

brdf

Surface reflectance type

decomp_type

Algorithm to use for fourier decomposition (NAI2/PCW)

quadrature_type

Quadrature type for RT streams (RadauQuad/GaussQuadHemisphere/GaussQuadFullSphere)

architecture

Architecture to use for calculations (CPU/GPU)

Model Type

RadiativeTransfer.vSmartMOM.vSmartMOM_Model — Type

struct vSmartMOM_Model

A struct which holds all derived model parameters (including any computations)

Fields

params

Struct with all individual parameters

aerosol_optics

Truncated aerosol optics

greek_rayleigh

Greek coefs in Rayleigh calculations

quad_points

Quadrature points/weights, etc

τ_abs

Array to hold cross-sections over entire atmospheric profile

τRayl

Rayleigh optical thickness

τAer

Aerosol optical thickness

obs_geom

Observational Geometry (includes sza, vza, vaz)

profile

Atmospheric profile to use

Surface Types

RadiativeTransfer.vSmartMOM.AbstractSurfaceType — Type

Abstract Type for Surface Types

RadiativeTransfer.vSmartMOM.LambertianSurfaceScalar — Type

Lambertian Surface (scalar per band)

RadiativeTransfer.vSmartMOM.LambertianSurfaceSpectrum — Type

Defined as Array (has to have the same length as the band!)

RadiativeTransfer.vSmartMOM.LambertianSurfacePolyFit — Type

Defined by polynomial terms as function of λ (degree derived from length of a_coeff)

Quadrature Types

RadiativeTransfer.vSmartMOM.AbstractQuadratureType — Type

Quadrature Types for RT streams

RadiativeTransfer.vSmartMOM.RadauQuad — Type

struct RadauQuad

Use Gauss Radau quadrature scheme, which includes the SZA as point (see Sanghavi vSmartMOM)

RadiativeTransfer.vSmartMOM.GaussQuadHemisphere — Type

struct GaussQuadHemisphere

Use Gauss quadrature scheme, define interval [-1,1] within an hemisphere (90⁰), repeat for both

RadiativeTransfer.vSmartMOM.GaussQuadFullSphere — Type

struct GaussQuadFullSphere

Use Gauss quadrature scheme, define interval [-1,1] for full sphere (180⁰), take half of it (less points near horizon compared to GaussQuadHemisphere)

Atmospheric Profile Type

RadiativeTransfer.vSmartMOM.AtmosphericProfile — Type

Struct for an atmospheric profile