RadiativeTransfer Module Methods & Types

Defining Parameters for RT Simulation

RadiativeTransfer.vSmartMOM.parameters_from_yaml — Function

Given a path to a YAML parameters file, load all the parameters into a vSmartMOM_Parameters struct

RadiativeTransfer.vSmartMOM.default_parameters — Function

Generate default set of parameters for Radiative Transfer calculations (from ModelParameters/)

Using the Parameters Object to Create a Model

RadiativeTransfer.vSmartMOM.model_from_parameters — Function

Take the parameters specified in the vSmartMOMParameters struct, and calculate derived attributes into a vSmartMOMModel

Performing the RT Simulation

RadiativeTransfer.vSmartMOM.rt_run — Function

rt_run(pol_type, obs_geom::ObsGeometry, τ_rayl, τ_aer, quad_points::QuadPoints, max_m, aerosol_optics, greek_rayleigh, τ_abs, brdf, architecture::AbstractArchitecture)

Perform Radiative Transfer calculations using given parameters

rt_run(model::vSmartMOM_Model, i_band::Integer = -1)

Perform Radiative Transfer calculations using parameters passed in through the vSmartMOM_Model struct

Types

Parameters Type

RadiativeTransfer.vSmartMOM.vSmartMOM_Parameters — Type

struct vSmartMOM_Parameters

A struct which holds all initial model parameters (before any computations)

Fields

spec_bands

Spectral bands (nBand)

brdf

Surface (Bidirectional Reflectance Distribution Function)

quadrature_type

Quadrature type for RT streams (RadauQuad/GaussQuadHemisphere/GaussQuadFullSphere)

polarization_type

Type of polarization (I/IQ/IQU/IQUV)

max_m

Hard cutoff for maximum number of Fourier moments to loop over

Δ_angle

Exclusion angle for forward peak [deg]

l_trunc

Truncation length for legendre terms (scalar for now, can do nBand later)

depol

Depolarization factor

float_type

Float type to use in the RT (Float64/Float32)

architecture

Architecture to use for calculations (CPU/GPU)

sza

Solar zenith angle [deg]

vza

Viewing zenith angles [deg]

vaz

Viewing azimuthal angles [deg]

obs_alt

Altitude of observer [Pa]

T

Temperature Profile [K]

p

Pressure Profile [hPa]

q

Specific humidity profile

profile_reduction_n

Length of profile reduction

absorption_params

Optional struct that holds all absorption-related parameters

scattering_params

Optional struct that holds all aerosol scattering-related parameters

Model Type

RadiativeTransfer.vSmartMOM.vSmartMOM_Model — Type

struct vSmartMOM_Model

A struct which holds all derived model parameters (including any computations)

Fields

params

Struct with all individual parameters

aerosol_optics

Truncated aerosol optics

greek_rayleigh

Greek coefs in Rayleigh calculations

quad_points

Quadrature points/weights, etc

τ_abs

Array to hold cross-sections over entire atmospheric profile

τ_rayl

Rayleigh optical thickness

τ_aer

Aerosol optical thickness

obs_geom

Observational Geometry (includes sza, vza, vaz)

profile

Atmospheric profile to use

Surface Types

RadiativeTransfer.vSmartMOM.AbstractSurfaceType — Type

Abstract Type for Surface Types

RadiativeTransfer.vSmartMOM.LambertianSurfaceScalar — Type

Lambertian Surface (scalar per band)

RadiativeTransfer.vSmartMOM.LambertianSurfaceSpectrum — Type

Defined as Array (has to have the same length as the band!)

RadiativeTransfer.vSmartMOM.LambertianSurfacePolyFit — Type

Defined by polynomial terms as function of λ (degree derived from length of a_coeff)

Quadrature Types

RadiativeTransfer.vSmartMOM.AbstractQuadratureType — Type

Quadrature Types for RT streams

RadiativeTransfer.vSmartMOM.RadauQuad — Type

struct RadauQuad

Use Gauss Radau quadrature scheme, which includes the SZA as point (see Sanghavi vSmartMOM)

RadiativeTransfer.vSmartMOM.GaussQuadHemisphere — Type

struct GaussQuadHemisphere

Use Gauss quadrature scheme, define interval [-1,1] within an hemisphere (90⁰), repeat for both

RadiativeTransfer.vSmartMOM.GaussQuadFullSphere — Type

struct GaussQuadFullSphere

Use Gauss quadrature scheme, define interval [-1,1] for full sphere (180⁰), take half of it (less points near horizon compared to GaussQuadHemisphere)

Atmospheric Profile Type

RadiativeTransfer.vSmartMOM.AtmosphericProfile — Type

Struct for an atmospheric profile