Scattering Module

Introduction

This module enables scattering phase-function calculation of atmospheric aerosols with different size distributions, incident wavelengths, and refractive indices. It can perform the calculation using either the Siewert NAI-2 or Domke PCW methods (Suniti Sanghavi 2014).

You can calculate a phase function in three steps:

Use make_univariate_aerosol to create an aerosol with selected distribution and properties

Use make_mie_model to set up all calculation parameters

Use compute_aerosol_optical_properties to perform the phase-function calculation using the defined model settings

Example

using RadiativeTransfer
using RadiativeTransfer.Scattering
using Distributions
using BenchmarkTools

## 
## STEP 1: Create the Aerosol
## 

# Aerosol particle distribution and properties 
μ  = 0.3                # Log mean radius
σ  = 6.82               # Log stddev of radius
r_max = 30.0            # Maximum radius
nquad_radius = 2500     # Number of quadrature points for integrating of size dist.
nᵣ = 1.3                # Real part of refractive index
nᵢ = 0.0                # Imag part of refractive index

size_distribution = LogNormal(log(μ), log(σ))

# Create the aerosol
aero = make_univariate_aerosol(size_distribution, r_max, nquad_radius, nᵣ, nᵢ)

## 
## STEP 2: Create the Mie Calculations model
## 

λ = 0.55   # Incident wavelength
polarization_type = Stokes_IQUV()
truncation_type = δBGE(10, 2)

# NAI2 Method
model_NAI2 = make_mie_model(NAI2(), aero, λ, polarization_type, truncation_type)

# PCW Method with saved/loaded Wigner tables
wigner_file_path = "PATH_TO_SAVED_WIGNER_MATRIX"
model_PCW = make_mie_model(PCW(), aero, λ, polarization_type, truncation_type, wigner_file_path)

# PCW Method with newly computed Wigner tables
wigner_A, wigner_B = compute_wigner_values(600) # Specify N_max
model_PCW_computed_wigner = make_mie_model(PCW(), aero, λ, polarization_type, truncation_type, wigner_A, wigner_B)

## 
## STEP 3: Perform the Mie Calculations
## 

aerosol_optics_NAI2 = compute_aerosol_optical_properties(model_NAI2);
aerosol_optics_PCW = compute_aerosol_optical_properties(model_PCW);

References

1. Domke, H., 1975. Fourier expansion of the phase matrix for Mie scattering. Zeitschrift für 
Meteorologie, 25(6), pp.357-361.

2. Siewert, C.E., 1982. On the phase matrix basic to the scattering of polarized light. 
Astronomy and Astrophysics, 109, p.195.

3. Sanghavi, S., 2014. Revisiting the Fourier expansion of Mie scattering matrices in 
generalized spherical functions. Journal of Quantitative Spectroscopy and Radiative Transfer, 
136, pp.16-27.

Creating the Aerosol

RadiativeTransfer.Scattering.make_univariate_aerosol — Function

make_univariate_aerosol(size_distribution::ContinuousUnivariateDistribution, r_max, nquad_radius::Int, nᵣ, nᵢ)

Convenience function to create a Univariate Aerosol, given the size parameters

Creating/Saving Wigner Matrices (if using PCW method)

RadiativeTransfer.Scattering.compute_wigner_values — Function

compute_wigner_values(m_max::Integer, n_max::Integer, l_max::Integer)

Compute the Wigner 3j values for all (m, n, l) combinations up to m/n/l max, for (m1, m2, m3) = (-1, 1, 0) (wignerA) and (-1, -1, 2) (wignerB)

compute_wigner_values(N_max)

Shorthand for computewignervalues(2Nmax + 1, Nmax + 1, 2N_max + 1)

RadiativeTransfer.Scattering.save_wigner_values — Function

save_wigner_values(filepath, wigner_A, wigner_B)

Save the Wigner A and Wigner B matrices at the given filepath

RadiativeTransfer.Scattering.load_wigner_values — Function

load_wigner_values(filename)

Load the Wigner A and Wigner B matrices from the given filepath

Defining the Mie Computation Model

RadiativeTransfer.Scattering.make_mie_model — Function

make_mie_model(computation_type::NAI2, aerosol::AbstractAerosolType, λ::Real, polarization::AbstractPolarizationType, truncation_type::AbstractTruncationType)

Convenience function to create Mie Model with NAI2 computation type

make_mie_model(computation_type::PCW, aerosol::AbstractAerosolType, λ::Real, polarization::AbstractPolarizationType, truncation_type::AbstractTruncationType, wigner_filepath::String)

Convenience function to load Wigner matrices from file and create Mie Model with PCW computation type

make_mie_model(computation_type::PCW, aerosol::AbstractAerosolType, λ::Real, polarization::AbstractPolarizationType, truncation_type::AbstractTruncationType, wigner_A, wigner_B)

Convenience function to take in Wigner matrices and create Mie Model with PCW computation type

Computing Aerosol Optical Properties

RadiativeTransfer.Scattering.compute_aerosol_optical_properties — Function

compute_aerosol_optical_properties(model::MieModel{FDT}) where FDT<:NAI2

Reference: Suniti Sanghavi 2014, https://doi.org/10.1016/j.jqsrt.2013.12.015

Compute the aerosol optical properties using the Siewert-NAI2 method Input: MieModel, holding all computation and aerosol properties Output: AerosolOptics, holding all Greek coefficients and Cross-Sectional information

compute_aerosol_optical_properties(model::MieModel{FDT}) where FDT<:PCW

Reference: Suniti Sanghavi 2014, https://doi.org/10.1016/j.jqsrt.2013.12.015

Compute the aerosol optical properties using the Domke-PCW method Input: MieModel, holding all computation and aerosol properties Output: AerosolOptics, holding all Greek coefficients and Cross-Sectional information

compute_aerosol_optical_properties(model::MieModel{FDT})

Reference: Suniti Sanghavi 2014, https://doi.org/10.1016/j.jqsrt.2013.12.015

This function enables user to specify whether to perform auto-differentiation (using either computation type) Input: MieModel, holding all computation and aerosol properties & autodiff flag (whether to perform auto-differentiation) Output: AerosolOptics, holding all Greek coefficients and Cross-Sectional information

Types

Aerosol Types

RadiativeTransfer.Scattering.UnivariateAerosol — Type

struct UnivariateAerosol{FT}

A struct which provides all univariate aerosol parameters needed for Mie computation

Fields

size_distribution

Univariate size distribution

r_max

Maximum radius [μm]

nquad_radius

Number of quadrature points for integration over size distribution

nᵣ

Real part of refractive Index

nᵢ

Imaginary part of refractive Index

Fourier Decomposition Computation Types

RadiativeTransfer.Scattering.NAI2 — Type

type NAI2

Perform Siewart's numerical integration method, NAI-2, to compute aerosol phase function decomposition. See: http://adsabs.harvard.edu/full/1982A%26A...109..195S

RadiativeTransfer.Scattering.PCW — Type

type PCW

Perform Domke's Precomputed Wigner Symbols method, PCW, to compute aerosol phase function decomposition. See: http://adsabs.harvard.edu/full/1984A%26A...131..237D

Polarization Types

RadiativeTransfer.Scattering.Stokes_IQUV — Type

struct Stokes_IQUV{FT<:AbstractFloat}

A struct which defines full Stokes Vector ([I,Q,U,V]) RT code

Fields

n

Number of Stokes components (int)

D

Vector of length n for ... (see eq in Sanghavi )

I₀

Incoming Stokes vector for scalar only

RadiativeTransfer.Scattering.Stokes_IQU — Type

struct Stokes_IQU{FT<:AbstractFloat}

A struct which defines Stokes Vector ([I,Q,U]) RT code

Fields

n

Number of Stokes components (int)

D

Vector of length n for ... (see eq in Sanghavi )

I₀

Incoming Stokes vector for scalar only

RadiativeTransfer.Scattering.Stokes_I — Type

struct Stokes_I{FT<:AbstractFloat}

A struct which define scalar I only RT code

Fields

n

Number of Stokes components (int)

D

Vector of length n for ... (see eq in Sanghavi )

I₀

Incoming Stokes vector for scalar only

Truncation Types

RadiativeTransfer.Scattering.δBGE — Type

type δBGE{FT} <: AbstractTruncationType

Fields

l_max

Trunction length for legendre terms

Δ_angle

Exclusion angle for forward peak (in fitting procedure) [degrees]

Mie Computation Model Type

RadiativeTransfer.Scattering.MieModel — Type

type MieModel

Model to hold all Mie computation details for NAI2 and PCW

Fields

computation_type

aerosol

λ

polarization_type

truncation_type

wigner_A

wigner_B

Output Aerosol Optics Types

RadiativeTransfer.Scattering.GreekCoefs — Type

struct GreekCoefs{FT}

A struct which holds all Greek coefficient lists (over l) in one object. See eq 16 in Sanghavi 2014 for details.

Fields

α

Greek matrix coefficient α, is in B[2,2]

β

Greek matrix coefficient β, is in B1,1

γ

Greek matrix coefficient γ, is in B[2,1],B[1,2]

δ

Greek matrix coefficient δ, is in B[4,4]

ϵ

Greek matrix coefficient ϵ, is in B[3,4] and - in B[4,3]

ζ

Greek matrix coefficient ζ, is in B[3,3]

RadiativeTransfer.Scattering.AerosolOptics — Type

struct AerosolOptics

A struct which holds all computed aerosol optics

Fields

greek_coefs

Greek matrix

ω̃

Single Scattering Albedo

k

Extinction cross-section

fᵗ

Truncation factor

derivs

Derivatives